Group adjacent events with the same type into one query
Summary:
For a write intensive table, we are currently not able to keep up the pace because we replay event one at a time, and the time is wasted in back and forth between client and mysql server. We also see a high cpu utilization during catchup for pure INSERT catchup because of this.
This change will try group the events into the one query as many as possible ( but also configurable )

Reviewed By: BMDan

Differential Revision: D4737086

fbshipit-source-id: 868ae2516448da4a2017e299a2faad78a597e5b5