Dump data after checksum to give better insights
Summary:
More aggressive version of checksum dump. `--detailed-checksum-mismatch` is good, but it will open a long transaction for both new and old table checksum, and lots of replay are issued inside the same transaction. This may become a performance issue under some workload.
So I implemented another version, which sacrifice the disk space for performance. This approach will dump the raw data onto local disk while we are calculating the checksum, which means we will use 3x the size of the table. By doing this we now have all the data from one snapshot on the disk. If there's a checksum mismatch, we can simply just `vimdiff table.new.11 table.old.11` to figure out what cause the mismatch.

Reviewed By: nathacof

Differential Revision: D4953353

fbshipit-source-id: d6d753e18979156c4f7f30f0dc8a0522709d2969