AOSC should only kill slow queries affecting the table being altered
Summary:
When OSC wants to swap a table or add triggers to it, it needs to hold the table lock. Slow queries though can hold this lock and OSC would get stuck. To avoid that OSC kills any selects on that table, but this also kills fast selects and cause connection loss.
This diff changes the mechanic to only trigger kill mechanics if we are waiting for more than X seconds, X being 0.5 by default (500 millis)

Reviewed By: cenalulu

Differential Revision: D5678197

fbshipit-source-id: 9e31713c874544c23e6524bba936f99403a68fa7