Set default value for '--max-wait-for-slow-query'
When --max_wait-for-slow-query is not supplied in the osc_cli command
line, that argument will still be created by Argparse with value `None`,
so the CopyPayload default init value will now be applied, since the
argument will be present.

This patch sets the default value for `max_wait_for_slow_query` using
the `constant.MAX_WAIT_FOR_SLOW_QUERY`.

Signed-off-by: Amador Pahim <amador@pahim.org>