support long table names with 54+ chars
Summary:
Currently OSC will add 10 chars in front of table name as a prefix, but it doesn't work with table having more than 54 characters in name because MySQL only supports 64 chars as an identifier.
In order to support longer names while maintain the visibility, this fix will switch to shorter prefixes only when it is necessary.

Reviewed By: guokeno0

Differential Revision: D6550337

fbshipit-source-id: 60f358fe7f84619fca6169b34f1f2e7673acf2b1