added force index to replay_insert
Summary:
MyRocks running out of row locks on seemingly simple queries that should only need a handful, causing OSC to fail.

Those queries should only have a handful of row operations, not even close to the million row locks it seems to try to take.

This change is supposed to address this problem.

Differential Revision: D8521288

fbshipit-source-id: a45f5d299a36d84642c519cc8178610ee1bd908c