Handle Time in processlist being NULL for processes in Connect status
Summary:
in running osc in python3 at scale for what may be the first time, we are hit by py3's inability to compare `None > 0` whereas py2 would have happily evaluated that to `False`.

The Time columne of "show (full) processlist" output can be NULL / None if the connection is in the "Connect" state still, so this accounts for that and makes sure we are only ever comparing int's.

Reviewed By: marteen

Differential Revision: D12878273

fbshipit-source-id: ab6591413770fec9a0e5a50e8ade122ec86f7a13