fix the trx leakage when checksum is disabled
Summary: In the current implementation we wrongly keep the transaction open for replay. This negative impact gets increase when checksum is disable where the replay2go has to replay tons of changes since table load, which will possibly cause the attempt to fail with max_row_lock_limit in MyRocks or leave the hist_length being big for too long.

Reviewed By: marteen

Differential Revision: D17123388

fbshipit-source-id: 182f8f37e4daf0ef8e60d3ff5ad89c6dc20f3782