Fix access to block_create_no_primary_key
Summary: The block_create_no_primary_key system variable in FB MySQL 5.6 is currently global only, and is about to be changed to global/session scope. After that changing the global variable is not going to affect the current session which would break the application. The fix tries to change the session variable first, and if that fails it changes the global variable. This way the app can work with either version of the server.

Reviewed By: cenalulu

Differential Revision: D21729815

fbshipit-source-id: 24523ee2dcb5edbd25398c415579364179804d5a