Enable running OSC tests in any engine
Summary: In order to automatically test OSC for any engine in merlin, the tests we have in OSC today has hard coded engine in the sql files and would end up creating InnoDB tables in MyRocks instances. This shouldn't be a block for AOSC since AOSC would strip the Engine value out, but only for OSC. For now I'm just stripping out the ENGINE=InnoDB (except in one test case where it is testing wrong engine) so that the test can run for both InnoDB and MyRocks instances.

Reviewed By: atish2196

Differential Revision: D22325475

fbshipit-source-id: 819c14849b6e7a1b5dad41dbd5430070b20250a5