osc - fail when we record too many row deltas to ever catchup
Summary: If a table fills the __osc_chg table to the point of max(int) value we will cause dup key errors from writes to the original table.  Prevent that and roll back if we get close to this limit.  The other option, increasing to bigint, would technically address this but given our throughput limitations thanks double-writing, it's unlikely such a schema change would ever finish via OSC.

Reviewed By: adregner

Differential Revision: D25041384

fbshipit-source-id: 2c4a3acb6bc56f55d29bf35d3d2fc1be828b00c4