Fix false negeative in integration test lib
Summary:
There's a bug in current implementation of the test library where there's no assertion to make sure 'expected error' is raised. Or in other words if there's no exception at all, the test will be considered as pass which is a false negative.
The bug is fixed in this diff. Also there's a broken test which was missed because of this and got fixed here as well.

Reviewed By: marteen

Differential Revision: D25228332

fbshipit-source-id: 2d04efaebd164984606cc7a0aec03a0c6c10d1d3