osc - apply partition drift to new table before swap
Summary: If the new version of the table has out-of-date partitions (ie, an ALTER to add or drop partitions came through during the osc process) this will almost certainly cause a breakage on rbr (as now partition between instances have drifted).  This won't cover all edge cases (this could still race with another DDL process or cause data drift if data is written to different partitions between instances) but most of the time this should prevent schema drift.

Reviewed By: srinivasmohan

Differential Revision: D26801114

fbshipit-source-id: 985a4fe17796fc7ee02557806a1766109469c95a