Do not check IO thread status while stopping SQL thread
Summary:
When OSC decides whether to stop the SQL thread on an instance, it checks for the status of both IO and SQL threads. It proceeds to stop the SQL thread (and later on resume it) iff both the threads are running.
This does not make sense because we should not worry about IO thread status when we want to stop only the SQL thread

Reviewed By: marteen

Differential Revision: D27862130

fbshipit-source-id: c88f316d6fe5e63e1dd845a2a0383b6b3a45c98f