Fix assertRaise contexts
Summary:
While checking unrelated tests I've found a very nasty source of wrong-placed test confidence.

The min-reproducible code looks as follow
```
from unittest import TestCase

def just_raise() -> None:
    raise RuntimeError("A special token")

class TestClass(TestCase):
    def test_something(self) -> None:
        with self.assertRaises(RuntimeError) as ex:
            just_raise()

            self.assertIn("custom token", str(ex.exception))
```
If you don't spot the issue I would invite to check the last assertion ... we are ensuring that "custom token" is in the string representation of the exception ... isn't it wrong?

If you move the `self.assertIn` outside of the with context the test would fail as expected.

**Why does it happen?**
Well .... `just_raise` raised an exception ... so `self.assertIn` is never executed.

**Why is this EVIL?**
Just because reading the test I believe that I've got automation ensuring that `custom token` is present in the exception string ... but it is not :(

## What does this diff do?
I've tried to identify source of this evil within `fbcode/dba/` and fix them all

Reviewed By: vladfau

Differential Revision: D30513329

fbshipit-source-id: 47d3342afedf4fd15587f63dc1d85ce8f8ab4e0f