aosc create table parsing - handle tablespace attribute
Summary: Doing an ALTER on a mysql 8.0 server without file-per-table will cause the SHOW CREATE TABLE output to include the TABLESPACE `innodb_system` comment.  Adding parsing support for this, and in case the value is innodb_system just remove the value from the CREATE statement (on the assumption that perhaps this open-source code is used somewhere where tablespaces matter).

Reviewed By: MarcoMontaltoMonella

Differential Revision: D37582200

fbshipit-source-id: cfdef3e169a01328b9fd96e2136ec599d86a29aa