For batch osc, re-require connection
Summary:
Cleanup workload and copy workload uses different connections. If we have multiple databases to work on, and one takes > session_timeout time to finish, then cleanup connection would be idle for too long and hit `mysql server gone away` error. Here, making a change so that if we have multiple databases, re-require the connection every time.

We could also factor in how long is it between cleanup, and compare with session wait timeout, but that will involve more interface changes and I'm not sure if it's worth it.

Reviewed By: lth

Differential Revision: D45612144

fbshipit-source-id: 08eb5d5f3f9b6f1947bc0946d63fd1403158a3e7