Cleanup upon cancellation
Summary: Currently when dbexecd_osc_wrapper is cancelled it just receives the signal and exits. The cleanup is an important stage so we need to take care of that. Between sigterm and sigkill there is by default 90 secs which should be enough to do cleanups. If cleanup fails we would rely on the aosc_local_checksum detecting leftover triggers and fbar remediation. Also logged metrics on how long cleanup takes.

Reviewed By: george-reynya

Differential Revision: D45744018

fbshipit-source-id: e64e3aaf3a121d9f302cbd64c8ac0775ed3121be