Fix ordering columns in generated alter statement
Summary:
There are bugs in SchemaDiff that it can't generate alter table stmt to change the column order in such cases.
{P887917902}

The reason is that if we move continuous columns together, the SchemaDiff just move the first column of the moved "block".

Now we keep the stable columns at the beginning and the end.

1. Add the new columns into the old column list
2. find the first modified column
3. do not move all the columns before the first one
4. calculate the distance to the end for all the columns after the first. If the distance changes, move it.

Differential Revision: D51540952

fbshipit-source-id: 881c313bc11854d3d8610a9e4eeb67a2a00c513a