Add a switch to use old partition parser
Summary:
We still need the old partition paser in Partition Manager. In the future, we will rewrite that and let it fit the new parser.

Temporarily use regular expression to catch the raw partition string. In the future we will get the partition string from the AST tree. To make sure the qeury matches the regular expression, add the c-style comment prefix/suffix into the test cases.

Reviewed By: avandever

Differential Revision: D51925127

fbshipit-source-id: afab498d4f34a75ec2fd56936126a77413daa801