Use the switch to select old/new parser in dba codebase (except OSC)
Summary:
Prepare for new parser rollout

I don't actually change parse_create but use the switch to select the old parser in dba.

When the new parser is fully rollout, we can switch to the new parser by changing the default value in fb_create.

And add the type definition to parse_create

Reviewed By: preritj24

Differential Revision: D51960485

fbshipit-source-id: 0a0d42dfd06851f6f27604213d55f74c22caa182