Do not fall back to traditional osc when co-osc failures are genuine schema change failures
Summary:
As part of the rollout of co-osc, we had a fallback to traditional osc whenever there is a failure on one of the replicas. This had been done so as to reduce the chances of a total replicaset level failure and annoying the user as well as wasted efforts in running the schema change. With a partial success, we could replace the failed instances whenever required with a successful one and get the schema change op. over the line.

This however was not the long term idea for co-osc as we want an all or nothing schema change always, so no more fallback for replica failures.

Differential Revision: D52423148

fbshipit-source-id: 38a1c4856261f6316f1e2b0f7b0b1ff73600c380