osc: Introduce fb-only osc cli tool for dev/test
Summary:
This allows for fast local iteration on the OSC codebase itself with just a local mysql on your devserver. This is bascially a fork of the open source `osc_cli` but with the changes necessary to run it buckified. As an added bonus, this allows us to exercise fb internal code too.

I modified the db helper to make the username/password optional since when connecting to local instance via unix socket there's no need for any password. Username is still required though (root).

This is much lower dependency than `dbexecd_osc_wrapper` or `dbatool3 osc`, allowing for more targetted testing and benchmarking.

Reviewed By: preritj24

Differential Revision: D52790396

fbshipit-source-id: 19264e277f109c981dfd8645add9aa1f35209041