osc: Add test for `get_range_start_condition`
Summary:
Adding a unit test to guarantee the behavior of aforementioned function. This has some non-trivial logic and does not have a unit test encoding its behavior.

Added a tests/sql.py file to the unit test suite for this and future SQL-generating test cases, few (any?) of which are tested in isolation today, and could be.

Reviewed By: preritj24

Differential Revision: D52790395

fbshipit-source-id: 98bf7852690ac03e77f8063f54a191504d424cb0