osc: Introduce load benchmark utility
Summary:
Similar to the dump benchmark of the previous diff, this does just the load part, leaving the chunk files around for reuse. You just have to drop the "copy table" yourself each time.

This allows us to benchmark and iterate on just the load section of the OSC code, easily comparing potential optimizations, etc.

Reviewed By: preritj24

Differential Revision: D52851164

fbshipit-source-id: 07958d5457b197c24cf2097131d01613f57cbfbf