Add some dumps on mismatched co-osc checksums
Summary:
We sprinkle some replication lag checks in co-osc to verify that instances are not totally backed up. This works fine for most usecases but occasionally, there are some heavy write customers, for which replication lag can be in a few seconds. Here, we will have a false positive in checksum calculations if we check for 240seconds lag and not bother about intermittent lag in seconds.

To detect such cases in general as well as any anomalies around replication, add a way to dump the checksums.

Another fix clubbed in is to clear the state of gap ids(last_replayed_id) from previous loop in a batch OSC, because that can cause us to skip tracking some gap ids in the current run, causing false positives.

Reviewed By: sveiss

Differential Revision: D57677277

fbshipit-source-id: 6355dc4fd6e07ae5077bfc92ca637f46bdd880f4