Ignore engine change in dry run output
Summary:
This ask came from bizdb, who use dryrun as a way to detect any pending schema changes automatically. Since they are going through a myrocks migration, this causes false positives.

This is really a patch since we want to do a more comprehensive change which doesn't care for engines. An alternative is to clear the engine field in each schema but we will see how the user feedback is.

Differential Revision: D58540076

fbshipit-source-id: 546bc99a4bda85b453e6cebdd5d12e84b52c7d53