osc: native checksum
Summary:
This diff adds native checksum support to the full table checksum stage in OSC (AOSC integration will be separate).

Full table checksum takes ~40% of the time in OSC.

More context:
1. https://fb.workplace.com/groups/dbeng.oncall/permalink/26271987645756502/
2. https://fb.workplace.com/groups/fb.database.management/permalink/1634178790698816/

Since AOSC changes the schema, the plain `CHECKSUM TABLE` statement cannot be used as is since it applies to all columns in the table: we need to compare the columns apples-to-apples. For that, D56383480 extended the command in the server to allow specific column selection, which also comes in handy for skipping non-deterministic columns like FLOAT (though for now AOSC does not perform such filtering, unlike OLM).

Reviewed By: ankurrastogi09

Differential Revision: D60481371

fbshipit-source-id: bd7837475c45f5e643bb902ed77da527f0bba24a