osc: support column reordering
Summary:
D60608640 adds support in the server to honor the column list order given to the `CHECKSUM TABLE` statement.

This change integrates it into OSC and teaches OSC to sort the `self.old_column_list` before sending it to the server to support column reorderings in the after table. E.g. if a table had columns (a, b, c) and an AOSC changed them to (a, c, b). Without this change and D60608640, `CHECKSUM TABLE` would process the columns in their defined order, resulting in different checksums. That would prevent us from using the faster native checksum for such AOSCs, and we'd have to special case this detection.

Now we sort the columns lexicographically when generating the SQL statement, so we send (a, b, c) in both cases to the server, and MySQL (with D60608640), respects this order and processes the columns in the order given, not the table definition's.

With these two diffs, we can handle such cases without a fallback, allowing the optimization to have more coverage.

Differential Revision: D60616829

fbshipit-source-id: 7e24a8ef269412a69db0f3dc4f12ae28cee3f71c