Checking that table schema is unchanged before loading data
Summary:
During S240364 it was observed that a DDL statement like `TRUNCATE TABLE` which is not captured by OSC triggers can be executed against the original table while the schema migration is in process. This can lead to an inconsistency between the original and new tables as the missing DDL statements are not played against the new table.

To protect against such unwanted outcomes this change is recording the original table timestamp from `information_schema` before snapshot generation and compares it to one captured right before load. If the two do not match the workflow is aborted with the error code 155.

MySQL does not update table create timestamp by default upon a truncation. To help with detection of truncation statement runs we have added a new global variable (`update_table_create_timestamp_on_truncate`) that is turned on to instruct MySQL to update the table create timestamp every time it is truncated.

To disassociate OSC rollout from production rollout of the new global variable a check is added to the OSC so that it uses the variable only if it is defined.

Reviewed By: preritj24

Differential Revision: D56985779

fbshipit-source-id: 1cfcfd1311c6a61b9cbbc124c3c1e53688352a54