co-osc: native checksum
Summary:
This diff integrates native checksum into co-osc using the server extensions added in D56383480.

The basic outline of how it works is similar to SQL-based checksum in co-osc.

Most of the fun happens on the primary. We:
1. Do a catchup and ensure all secondaries catch up using the same "marker" technique SQL-based checksum uses (drop a table on the primary and wait for the drop to replicate on all secondaries).
1. Establish a fresh (TCP) connection to all replicas, including ourselves, and issue a `CHECKSUM TABLE` statement against the new table.
1. While that is running, issue a `CHECKSUM TABLE` statement against the old table using the default (UNIX, local) connection.
1. Wait for all results to come back.
1. Compare the old table checksum against the new table checksums.

Also refactored the Warm Storage and Co-OSC checks and added unit tests to harden the contract.

Reviewed By: preritj24

Differential Revision: D61173262

fbshipit-source-id: 449e84a7ffd3c8576ccea65fe3d34a2b166c31dc