Making OSC table create timestamp change retryable
Summary:
In D56985779 we added a new failure model to OSC to fail a schema change if the table create timestamp is modified in the middle of the workflow. This is intended to detect a race condition in which OSC can potentially coincide with a user initiated change to the  table schema. This caused failures in production for payments usecase which we could not triage.

This change does two things: (1) it improves the logging of the error condition so that we can additionally see what the observed timestamp is and (2) modifies the error classification from non-retryable to retryable. We observed that the timestamp change failure happens sporadically. Hence it makes sense to make it retryable in favor of increasing experienced success rate in the AOSC pipeline.

Differential Revision: D62893842

fbshipit-source-id: c9a8d5c66b77fbed8c1b4ec1a7d9373f23a19a15