Convert MySQL schema sql to an intermediate Thrift for easy post-processing
Summary:
This is not a conversion of CREATE TABLE SQL to a thrift struct, this is an intermediate representation of the sql
into a MySQLTable which still has the types faithfully represented as MySQL types. i.e. VARCHAR(50) is represented
as a thrift class VarCharTypeInfo but with an additional length field.

For each mysql schema domain and table, we will dump out this as a JSON file mainly for bidirectional conversion from
thrift back to MySQL.

The next step is to use this IR to dump out the thrift struct. This will be in a followup diff. That thrift struct will
try to use native thrift types which are superset types of MySQL

Reviewed By: preritj24

Differential Revision: D65411762

fbshipit-source-id: 63dc981779b9271268fae54435ff11fe4b307ef5